* move the mailbox list into separate component * normalize the titan mailboxes list
