+++ client/my-sites/email/email-management/home/email-plan-view.jsx
@@ -14,17 +14,27 @@ import wp from 'calypso/lib/wp';
 import { hasEmailForwards } from 'calypso/lib/domains/email-forwarding';
 import { hasGSuiteWithUs } from 'calypso/lib/gsuite';
 import { hasTitanMailWithUs } from 'calypso/lib/titan';
-import SectionHeader from 'calypso/components/section-header';
 import HeaderCake from 'calypso/components/header-cake';
 import VerticalNav from 'calypso/components/vertical-nav';
 import VerticalNavItem from 'calypso/components/vertical-nav/item';
 import EmailTypeIcon from 'calypso/my-sites/email/email-management/home/email-type-icon';
-import MaterialIcon from 'calypso/components/material-icon';
-import Badge from 'calypso/components/badge';
 import getGSuiteUsers from 'calypso/state/selectors/get-gsuite-users';
 import { getEmailForwards } from 'calypso/state/selectors/get-email-forwards';
 import QueryEmailForwards from 'calypso/components/data/query-email-forwards';
 import { emailManagement } from 'calypso/my-sites/email/paths';
+import EmailPlanMailboxesList from 'calypso/my-sites/email/email-management/home/email-plan-mailboxes-list';
+
+const normalizeTitanMailboxes = ( titanMailboxes ) => {
+	if ( ! titanMailboxes ) {
+		return [];
+	}
+	return titanMailboxes.map( ( mailbox ) => {
+		return {
+			email: mailbox.email,
+			isAdmin: mailbox.is_admin,
+		};
+	} );
+};
 
 class EmailPlanView extends React.Component {
 	state = {
@@ -79,40 +89,10 @@ class EmailPlanView extends React.Component {
 		const { domain, emails } = this.props;
 
 		if ( hasTitanMailWithUs( domain ) ) {
-			return this.state.titanMailboxes;
+			return normalizeTitanMailboxes( this.state.titanMailboxes );
 		}
 
 		return emails;
-	}
-
-	renderEmailsList() {
-		const { translate } = this.props;
-
-		const emails = this.getEmails();
-
-		if ( ! emails || emails.length < 1 ) {
-			return;
-		}
-
-		const emailsItems = emails.map( ( email ) => {
-			return (
-				<CompactCard
-					key={ `email-row-${ email.email }` }
-					className="email-plan-view__mailbox-list-item"
-				>
-					<MaterialIcon icon="email" />
-					<span>{ email.email }</span>
-					{ email.isAdmin && <Badge type="info">{ translate( 'Admin' ) }</Badge> }
-				</CompactCard>
-			);
-		} );
-
-		return (
-			<div className="email-plan-view__mailbox-list">
-				<SectionHeader label={ translate( 'Mailboxes' ) } />
-				{ emailsItems }
-			</div>
-		);
 	}
 
 	handleBack = () => {
@@ -138,7 +118,7 @@ class EmailPlanView extends React.Component {
 				</CompactCard>
 				<CompactCard>Expires: date [renew now] [auto renew]</CompactCard>
 
-				{ this.renderEmailsList() }
+				<EmailPlanMailboxesList emails={ this.getEmails() } />
 
 				<VerticalNav>
 					<VerticalNavItem path="a">Add new mailbox</VerticalNavItem>
@@ -188,8 +168,6 @@ export default connect( ( state, ownProps ) => {
 	} else if ( hasEmailForwards( ownProps.domain ) ) {
 		const emailForwards = getEmailForwards( state, ownProps.domain?.name ) ?? [];
 		emails = normalizeEmailForwardingAddresses( emailForwards );
-	} else if ( hasTitanMailWithUs( ownProps.domain ) ) {
-		emails = []; // filterEmailListByDomain( ownProps.titanMailboxes, ownProps.domain?.name );
 	}
 
 	return {
